Added confirmation to DeleteCommand
